// src/types/restaurant.ts
export interface Restaurant {
    UC_SEQ: number;
    MAIN_TITLE: string;
    GUGUN_NM: string;
    LAT: number;
    LNG: number;
    PLACE: string;
    TITLE: string;
    SUBTITLE: string;
    ADDR1: string;
    ADDR2: string;
    CNTCT_TEL: string;
    HOMEPAGE_URL: string;
    USAGE_DAY_WEEK_AND_TIME: string;
    RPRSNTV_MENU: string;
    MAIN_IMG_NORMAL: string | null;
    MAIN_IMG_THUMB: string | null;
    ITEMCNTNTS: string;
}